Encapsulate globals and add progress file
