+++ STATUS.md
@@ -1,0 +1,3 @@
+# Projektstatus
+
+Fortschritt bis zum Release: 90%
