+++ TODO-AGENTS.md
@@ -1,3 +1,3 @@
 # Offene Agenten-Aufgaben
 
-- Globale Zustände in eigene Module kapseln und wiederkehrende Utilities bündeln.
+Aktuell keine. Alle geplanten Verbesserungen sind umgesetzt.
